+++ src/dominion/models/characteristic/Characteristic.java
@@ -1,4 +1,4 @@
-package dominion.models;
+package dominion.models.characteristic;
 
 public interface Characteristic {
     public CharacteristicName getName();