+++ src/dominion/models/characteristic/CharacteristicMap.java
@@ -1,6 +1,8 @@
-package dominion.models;
+package dominion.models.characteristic;
 
 import java.util.Map;
+
+import dominion.models.CharacteristicValue;
 
 public interface CharacteristicMap extends Map<Characteristic, CharacteristicValue> {