+++ src/dominion/models/characteristic/CharacteristicName.java
@@ -1,0 +1,7 @@
+package dominion.models.characteristic;
+
+import dominion.models.Name;
+
+public interface CharacteristicName extends Name{
+
+}