+++ src/dominion/models/characteristic/Modifier.java
@@ -1,5 +1,4 @@
-package dominion.models;
-
+package dominion.models.characteristic;
 
 public interface Modifier extends CharacteristicMap{