+++ src/dominion/models/persona/BasicModifier.java
@@ -2,8 +2,8 @@ package dominion.models.persona;
 
 import java.util.HashMap;
 
-import dominion.models.Characteristic;
 import dominion.models.CharacteristicValue;
+import dominion.models.characteristic.Characteristic;
 
 public class BasicModifier extends HashMap<Characteristic, CharacteristicValue> implements PersonaModifier {