+++ src/dominion/models/persona/PersonaCharacteristic.java
@@ -1,6 +1,6 @@
 package dominion.models.persona;
 
-import dominion.models.Characteristic;
+import dominion.models.characteristic.Characteristic;
 
 public interface PersonaCharacteristic extends Characteristic{