+++ src/dominion/models/persona/PersonaCharacteristicMap.java
@@ -1,6 +1,6 @@
 package dominion.models.persona;
 
-import dominion.models.CharacteristicMap;
+import dominion.models.characteristic.CharacteristicMap;
 
 public interface PersonaCharacteristicMap extends CharacteristicMap{